analytics/requirements.txt
celery==5.3.1
redis==4.5.5
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6